% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-solutions.R
\docType{data}
\name{aoc15_solutions}
\alias{aoc15_solutions}
\title{Advent of Code 2015 solutions}
\format{An object of class \code{list} of length 50.}
\usage{
aoc15_solutions
}
\description{
Advent of Code 2015 solutions
}
\keyword{datasets}
