% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{count_parens}
\alias{count_parens_until_floor}
\title{Day 01: Not Quite Lisp}
\usage{
count_parens(xs)

count_parens_until_floor(string, target)
}
\arguments{
\item{xs}{a vector of strings of parentheses}

\item{string}{string of parentheses to step through}

\item{target}{desired floor}
}
\description{
\href{http://adventofcode.com/2015/day/1}{Not Quite Lisp}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
count_parens(")())())")
count_parens(c("(((", "(()(()(", "))((((("))
count_parens_until_floor("))(((((", 1)
}
