% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{measure_paper_for_boxes}
\alias{measure_ribbon_for_boxes}
\title{Day 02: I Was Told There Would Be No Math}
\usage{
measure_paper_for_boxes(xs)

measure_ribbon_for_boxes(xs)
}
\arguments{
\item{xs}{a character vector of box dimensions}
}
\description{
\href{http://adventofcode.com/2015/day/2}{I Was Told There Would Be No Math}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
measure_paper_for_boxes(c("2x3x4", "1x1x10"))
measure_ribbon_for_boxes(c("2x3x4", "1x1x10"))
}
