% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{trace_path}
\alias{trace_both_paths}
\title{Day 03: Perfectly Spherical Houses in a Vacuum}
\usage{
trace_path(string)

trace_both_paths(string)
}
\arguments{
\item{string}{a string of steps}
}
\description{
\href{http://adventofcode.com/2015/day/3}{Perfectly Spherical Houses in a Vacuum}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
trace_path("^v^v^v^v^v")
trace_both_paths("^v^v^v^v^v")
}
