% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{mine_advent_coins}
\alias{earns_advent_coin}
\alias{earns_super_advent_coin}
\title{Day 04: The Ideal Stocking Stuffer}
\usage{
mine_advent_coins(seed, start = 1L, criterion = earns_advent_coin)

earns_advent_coin(x)

earns_super_advent_coin(x)
}
\arguments{
\item{seed}{seed for the hashing algorithm}

\item{start}{number to start at}

\item{criterion}{rule to judge hashes}

\item{x}{string to check}
}
\description{
\href{http://adventofcode.com/2015/day/4}{The Ideal Stocking Stuffer}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
mine_advent_coins("abcdef", 600000)
}
