% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{check_for_nice_words}
\alias{contains_three_vowels}
\alias{contains_a_repeated_letter}
\alias{lacks_illegal_strings}
\alias{contains_repeated_pair}
\alias{contains_staggered_repeat}
\title{Day 05: Doesn't He Have Intern-Elves For This?}
\usage{
check_for_nice_words(strings, rule_list)

contains_three_vowels(xs)

contains_a_repeated_letter(xs)

lacks_illegal_strings(xs, blacklist = c("ab", "cd", "pq", "xy"))

contains_repeated_pair(xs)

contains_staggered_repeat(xs)
}
\arguments{
\item{blacklist}{a vector of illegal sequences}
}
\description{
\href{http://adventofcode.com/2015/day/5}{Doesn't He Have Intern-Elves For This?}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
strings <- c("ugknbfddgicrmopn", "aaa", "jchzalrnumimnmhp",
             "haegwjzuvuyypxyu", "dvszwmarrgswjxmb")
rule_list <- list(contains_three_vowels, contains_a_repeated_letter,
                  lacks_illegal_strings)

check_for_nice_words(strings, rule_list)
}
