% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\docType{data}
\name{day06}
\alias{day06}
\alias{initialize_light_grid}
\alias{perform_light_command}
\alias{light_rules}
\alias{light_rules_alt}
\title{Day 06: Probably a Fire Hazard}
\format{An object of class \code{environment} of length 3.}
\usage{
initialize_light_grid(rows = 1000L, cols = 1000L, start = 0L)

perform_light_command(grid, command_text, rule_env = light_rules)

light_rules

light_rules_alt
}
\arguments{
\item{rows, cols}{dimensions of light grid}

\item{start}{initalize value for lights}

\item{grid}{a grid of lights}

\item{command_text}{a vector with lighting commands}

\item{rule_env}{environment containing definitions for \code{turn_on_lights()},
\code{turn_off_lights()} and \code{toggle_lights()}. Defaults to \code{light_rules} which
are the definitions for Part 1. \code{light_rules_alt} are the definitions for
Part 2.}
}
\description{
\href{http://adventofcode.com/2015/day/6}{Probably a Fire Hazard}
}
\details{
\strong{Part One}

\strong{Part Two}
}
\examples{
grid <- initialize_light_grid(7, 5, 0)
grid

grid <- perform_light_command(grid, "turn on 0,0 through 4,4")
grid

commands <- c("turn off 0,1 through 4,1", "turn off 0,3 through 4,3")
grid <- perform_light_command(grid, commands)
grid

grid <- perform_light_command(grid, "toggle 0,0 through 6,4")
grid
}
\keyword{datasets}
